feat: adding svg
