+++ app/about-us/WhatWeDo.jsx
@@ -1,0 +1,112 @@
+import React from "react";
+
+const WhatWeDo = () => {
+  return (
+    <section className="py-20">
+      <div className="px-4 sm:px-10 md:px-12 lg:px-5">
+        <div className="flex flex-wrap mx-[-15px] xl:mx-[-35px] lg:mx-[-20px] mt-[-50px] items-center">
+          <div className="xl:w-7/12 lg:w-7/12 w-full flex-[0_0_auto] xl:px-[35px] lg:px-[20px] px-[15px] max-w-full xl:!order-2 lg:!order-2 mt-[50px]">
+            <div className="flex flex-wrap mx-[-15px] xl:mx-[-12.5px] lg:mx-[-12.5px] md:mx-[-12.5px] mt-[-25px]">
+              <div className="xl:w-5/12 lg:w-5/12 md:w-5/12 w-full flex-[0_0_auto] xl:px-[12.5px] lg:px-[12.5px] md:px-[12.5px] px-[15px] max-w-full xl:!ml-[8.33333333%] lg:!ml-[8.33333333%] md:!ml-[8.33333333%] !self-end mt-[25px]">
+                <div className="rounded-xl !bg-[#fef3e4]">
+                  <div className="p-[40px]">
+                    <img src="/svg/seo.svg" className="w-11 mb-3" alt="icon" />
+                    <h4>SEO Services</h4>
+                    <p className="!mb-0">
+                      Nulla vitae elit libero, a pharetra augue. Donec id elit
+                      non mi porta.
+                    </p>
+                  </div>
+                </div>
+              </div>
+
+              <div className="xl:w-6/12 lg:w-6/12 md:w-6/12 w-full flex-[0_0_auto] xl:px-[12.5px] lg:px-[12.5px] md:px-[12.5px] px-[15px] max-w-full !self-end mt-[25px]">
+                <div className="rounded-xl !bg-[#fae6e7]">
+                  <div className="p-[40px]">
+                    <img src="/svg/tab.svg" className="w-11 mb-3" alt="icon" />
+                    <h4>Web Design</h4>
+                    <p className="!mb-0">
+                      Nulla vitae elit libero, a pharetra augue. Donec id elit
+                      non mi porta.
+                    </p>
+                  </div>
+                </div>
+              </div>
+
+              <div className="xl:w-5/12 lg:w-5/12 md:w-5/12 w-full flex-[0_0_auto] xl:px-[12.5px] lg:px-[12.5px] md:px-[12.5px] px-[15px] max-w-full mt-[25px]">
+                <div className="rounded-xl !bg-[#eaf3ef]">
+                  <div className=" p-[40px]">
+                    <img
+                      src="/svg/chat-2.svg"
+                      className="w-11 mb-3"
+                      alt="icon"
+                    />
+                    <h4>Social Engagement</h4>
+                    <p className="!mb-0">
+                      Nulla vitae elit libero, a pharetra augue.
+                    </p>
+                  </div>
+                </div>
+              </div>
+
+              <div className="xl:w-6/12 lg:w-6/12 md:w-6/12 w-full flex-[0_0_auto] xl:px-[12.5px] lg:px-[12.5px] md:px-[12.5px] px-[15px] max-w-full !self-start mt-[25px]">
+                <div className="rounded-xl !bg-[#e0e9fa]">
+                  <div className=" p-[40px]">
+                    <img src="/svg/mics.svg" className="w-11 mb-3" alt="icon" />
+                    <h4>Content Marketing</h4>
+                    <p className="!mb-0">
+                      Nulla vitae elit libero, a pharetra augue. Donec id elit
+                      non mi porta gravida at eget.
+                    </p>
+                  </div>
+                </div>
+              </div>
+            </div>
+          </div>
+
+          <div className="xl:w-5/12 lg:w-5/12 w-full flex-[0_0_auto] xl:px-[35px] lg:px-[20px] px-[15px] mt-[50px] max-w-full relative">
+            <h2 className="!text-[.75rem] uppercase text-[#aab0bc] !mb-3 tracking-[0.02rem] leading-[1.35]">
+              What We Do?
+            </h2>
+            <h3 className="text-lg font-semibold lg:text-4xl mb-3 !leading-[1.3] text-gray-700">
+              The service we offer is specifically designed to meet your needs.
+            </h3>
+            <p className="text-[21px] leading-relaxed font-medium mb-3 text-gray-500">
+              We are a digital and branding company that believes in the power
+              of creative strategy and along with great design.
+            </p>
+
+            <svg
+              xmlns="http://www.w3.org/2000/svg"
+              width="206"
+              height="214"
+              viewBox="0 0 206 214"
+              fill="none"
+              className="absolute top-0 -translate-y-40 left-80 rotate-12"
+            >
+              <g clipPath="url(#clip0_106_19)">
+                <path
+                  d="M32.63 128.891a53.114 53.114 0 013.285-3.238.832.832 0 011.187.066.834.834 0 01-.07 1.188 50.137 50.137 0 00-4.608 4.726.83.83 0 01-1.182.093c-.343-.319-.392-.823-.088-1.182a72.21 72.21 0 011.476-1.653zm-5.447 7.003a.839.839 0 011.07-.135c.388.238.5.771.245 1.144a53.747 53.747 0 00-3.23 5.765.84.84 0 11-1.51-.733c1.034-2.094 2.134-4.095 3.334-5.943l.106-.114-.015.016zm13.963-14.328l.122-.131a50.986 50.986 0 015.853-3.478.842.842 0 011.119.369.846.846 0 01-.372 1.12 47.863 47.863 0 00-5.667 3.376c-.386.252-.91.161-1.161-.224a.841.841 0 01.09-1.047l.016.015zm-19.835 26.373a.815.815 0 01.893-.208c.433.158.67.623.511 1.056a58.115 58.115 0 00-1.819 6.343c-.11.447-.543.748-.989.638-.446-.109-.73-.557-.635-.987.478-2.183 1.128-4.389 1.876-6.535a.896.896 0 01.18-.292l-.017-.015zm31.777-32.437a.747.747 0 01.325-.218 54.822 54.822 0 015.648-1.621c.298-.059.612-.134.91-.193a.825.825 0 01.986.643.83.83 0 01-.645.988c-.298.059-.58.133-.878.192-1.85.418-3.681.945-5.477 1.567a.864.864 0 01-1.076-.512.84.84 0 01.176-.845l.031-.001zm53.965-57.967c1.051-1.13 2.135-2.23 3.205-3.282a.843.843 0 011.185.002.843.843 0 01-.006 1.186 93.784 93.784 0 00-4.59 4.805.825.825 0 01-1.168.044.83.83 0 01-.057-1.183c.488-.524.974-1.08 1.477-1.62l-.046.048zm-5.739 6.761c.289-.31.793-.362 1.119-.09a.832.832 0 01.151 1.18 124.248 124.248 0 00-3.938 5.381.852.852 0 01-1.177.219.85.85 0 01-.215-1.178 127.057 127.057 0 013.983-5.461l.045-.05h.032zm13.926-14.484l.076-.082a78.22 78.22 0 015.422-4.062c.386-.252.911-.161 1.161.224a.842.842 0 01-.228 1.163 68.687 68.687 0 00-5.314 3.979.828.828 0 01-1.173-.114.838.838 0 01.041-1.092l.015-.016zM93.685 75.349a.837.837 0 011.053-.15.835.835 0 01.262 1.16c-.809 1.294-1.54 2.507-2.215 3.592a164 164 0 00-1.284 2.135.839.839 0 01-1.144.28.836.836 0 01-.277-1.143c.433-.727.882-1.44 1.285-2.135.674-1.084 1.406-2.297 2.23-3.609l.106-.115-.016-.015zm32.657-33.055s.121-.131.215-.166a70.072 70.072 0 016.108-2.967.858.858 0 011.106.448c.188.42-.032.918-.452 1.108a70.708 70.708 0 00-5.968 2.882.842.842 0 01-1.133-.32.839.839 0 01.124-.985v0zM86.864 86.909a.814.814 0 011.005-.165c.403.222.578.72.355 1.124a65.65 65.65 0 00-2.827 5.987.847.847 0 01-1.09.452.843.843 0 01-.449-1.089 71.152 71.152 0 012.901-6.132c.046-.049.09-.13.136-.179l-.031.002zM18.43 161.024a.862.862 0 01.702-.264.83.83 0 01.759.92 48.471 48.471 0 00-.183 6.581.826.826 0 01-.808.869.82.82 0 01-.865-.805 51.832 51.832 0 01.207-6.804.899.899 0 01.22-.498l-.032.001zm47.775-48.233a.782.782 0 01.545-.258 54.513 54.513 0 016.806-.116c.46.03.805.412.805.87a.87.87 0 01-.873.808 54.181 54.181 0 00-6.601.124.84.84 0 01-.896-.772.814.814 0 01.23-.641l-.016-.015zm72.472-75.707a1 1 0 01.371-.235 71.284 71.284 0 016.501-1.654l.078-.019a.823.823 0 01.971.659.83.83 0 01-.662.973l-.063.002c-2.134.43-4.28.97-6.344 1.616a.834.834 0 01-1.046-.544.859.859 0 01.177-.813l.017.015zM81.794 99.302c.213-.23.525-.336.829-.237.43.126.7.59.573 1.022a48.05 48.05 0 00-1.373 6.453.832.832 0 01-.94.715.825.825 0 01-.711-.937 52.434 52.434 0 011.413-6.66.736.736 0 01.193-.34l.016-.016zm-62.87 75.074a.76.76 0 01.481-.255c.456-.065.893.219.972.69a62.35 62.35 0 00.96 4.467.84.84 0 01-.596 1.035.834.834 0 01-1.031-.593 46.384 46.384 0 01-.998-4.625.827.827 0 01.212-.719v0zm132.872-139.94a.86.86 0 01.528-.273 74.525 74.525 0 016.768-.336.84.84 0 01.834.837.846.846 0 01-.84.838 71.067 71.067 0 00-6.609.33c-.473.05-.864-.283-.912-.755a.869.869 0 01.215-.656l.016.015zm-71.915 78.088a.881.881 0 01.623-.276.84.84 0 01.835.837c.002.063-.012.111-.01.175 1.893.354 3.773.789 5.592 1.32.448.141.7.59.558 1.038-.143.448-.592.702-1.04.561a61.21 61.21 0 00-5.099-1.229c.055 1.579.157 3.171.306 4.714a.849.849 0 01-.743.929.841.841 0 01-.926-.739 81.636 81.636 0 01-.338-5.203l-.002-.063c-.244-.196-.38-.475-.313-.81a.966.966 0 01.314-.533c-.003-.063.011-.111.01-.174a.78.78 0 01.217-.562l.016.015zm85.235-78.136a.86.86 0 01.702-.263c2.219.2 4.474.525 6.701.946a.833.833 0 01.667.986.844.844 0 01-.988.67 73.107 73.107 0 00-6.543-.921c-.459-.03-.79-.46-.759-.92a.898.898 0 01.22-.498v0zm-83.874 91.408a.741.741 0 01.434-.238.85.85 0 011.001.626c.504 2.209 1.114 4.335 1.863 6.314a.831.831 0 01-.485 1.077.827.827 0 01-1.075-.48 51.6 51.6 0 01-1.933-6.533c-.073-.282.028-.554.211-.751l-.016-.015zm11.075-8.734a.84.84 0 01.941-.194c.786.334 1.573.699 2.329 1.065 1.27.632 2.51 1.296 3.72 1.993a.852.852 0 01.292 1.158.856.856 0 01-1.159.297 64.38 64.38 0 00-3.607-1.949 40.958 40.958 0 00-2.264-1.036.828.828 0 01-.434-1.106 1.13 1.13 0 01.166-.243l.016.015zm85.953-80.218c.213-.23.525-.336.829-.237a80.794 80.794 0 016.441 2.079.835.835 0 01.495 1.071.84.84 0 01-1.073.5 77.226 77.226 0 00-6.297-2.037.87.87 0 01-.589-1.037.731.731 0 01.194-.339v0zM85.956 138.276s.107-.115.2-.15a.837.837 0 011.15.304c1.141 1.948 2.42 3.717 3.831 5.244a.826.826 0 01-.054 1.172.825.825 0 01-1.171-.051 31.878 31.878 0 01-4.048-5.536.858.858 0 01.108-.999l-.015.016zm17.832-14.427c.289-.312.793-.362 1.119-.09.958.738 1.853 1.51 2.686 2.284a31.584 31.584 0 012.31 2.362.834.834 0 01-.069 1.188.833.833 0 01-1.188-.065 35.77 35.77 0 00-2.211-2.24 35.579 35.579 0 00-2.588-2.193.848.848 0 01-.152-1.18l.046-.049.047-.017zm87.032-82.361a.839.839 0 01.973-.195 93.668 93.668 0 016.017 3.09c.403.222.531.739.308 1.142-.223.404-.74.534-1.142.312a76.982 76.982 0 00-5.889-3.032c-.418-.205-.609-.688-.402-1.107.045-.08.089-.161.15-.227l-.015.017zM95 148a.84.84 0 011.07-.136c1.875 1.194 3.897 2.097 5.842 2.592a.84.84 0 01.605 1.02.848.848 0 01-1.022.608c-2.12-.536-4.32-1.512-6.34-2.81-.387-.238-.5-.771-.261-1.159.03-.033.06-.066.107-.115v0zm17.527-14.1s.137-.147.231-.182a.843.843 0 011.119.368 20.268 20.268 0 011.58 4.097c.219.845.359 1.709.406 2.608.032.457-.334.85-.807.868-.457.033-.849-.331-.866-.805a10.828 10.828 0 00-.363-2.262 18.197 18.197 0 00-1.458-3.769c-.153-.326-.103-.707.125-.953l.033.03zm-4.806 16.825c.092-.098.231-.183.371-.235a9.119 9.119 0 002.301-1.099 8.375 8.375 0 002.377-2.556.84.84 0 011.144-.281.837.837 0 01.276 1.144 10.274 10.274 0 01-2.864 3.08 10.227 10.227 0 01-2.766 1.338.809.809 0 01-1.031-.562.854.854 0 01.177-.813l.015-.016z"
+                  fill="#605DBA"
+                  stroke="#605DBA"
+                  strokeMiterlimit="10"
+                />
+              </g>
+              <defs>
+                <clipPath id="clip0_106_19">
+                  <path
+                    fill="#fff"
+                    transform="rotate(132.92 91.877 70.178)"
+                    d="M0 0h222.224v74.429H0z"
+                  />
+                </clipPath>
+              </defs>
+            </svg>
+          </div>
+        </div>
+      </div>
+    </section>
+  );
+};
+
+export default WhatWeDo;
